requirements.txt
ipykernel==6.30.1
matplotlib==3.10.5
pilkit==3.0
pillow==11.3.0
scikit-learn==1.7.1
torch==2.8.0
torchvision==0.23.0
